packages

+++ controllers/packageController.js
@@ -1,3 +1,4 @@
+const OwnerPackage = require('../models/ownerPackageModel')
 const Package = require('../models/packageModel')
 
 
@@ -27,8 +28,18 @@ const updatePackage = async(req,res) => {
 
 const getPackages = async(req,res) => {
     try {
-        const package = await Package.find()
-        res.status(200).send(package)
+        const packages = await Package.find()
+        res.status(200).send(packages)
+    } catch (error) {
+        console.log(error)
+        res.status(500).send({error})
+    }
+  }
+
+const getActivePackages = async(req,res) => {
+    try {
+        const packagse = await Package.find({status : true})
+        res.status(200).send(packagse)
     } catch (error) {
         console.log(error)
         res.status(500).send({error})
@@ -38,8 +49,13 @@ const getPackages = async(req,res) => {
 const deletePackage = async(req,res) => {
     try {
         const _id = req.query.id
-        await Package.deleteOne({_id})
-        res.status(200).json({message : "Package removed."})
+        const isPack = await OwnerPackage.find({pack : _id})
+        if (isPack.length > 0) {
+            res.status(400).json()
+        } else {
+            await Package.deleteOne({_id})
+            res.status(200).json({message : "Package removed."})
+        }
     } catch (error) {
         res.status(500).send(error)
     }
@@ -51,5 +67,6 @@ module.exports = {
     getPackages,
     createPackage,
     updatePackage,
-    deletePackage
+    deletePackage,
+    getActivePackages
 }
